Add DB object #639799
Also sets us up to use MySQL. SQLite is a pain, and doesn't have as many
features as MySQL.
This pull is untested.
The issue with MySQL is it took a lot of setup to work around Cal Poly
blocking students from doing anything useful. It may only work when
connected to Cal Poly's network.
I'll keep working on getting around that issue and test it out this
week.